Minor fixes; tutorial updates.
git-svn-id: https://svn.apache.org/repos/asf/pivot/trunk@1003111 13f79535-47bb-0310-9956-ffa450edef68

+++ wtk/src/org/apache/pivot/wtk/skin/TextAreaSkinParagraphView.java
@@ -283,7 +283,7 @@ class TextAreaSkinParagraphView implements Visual, TextArea.ParagraphListener {
         LineMetrics lm = font.getLineMetrics("", fontRenderContext);
         float rowHeight = lm.getAscent() + lm.getDescent();
 
-        int i = (int)Math.floor((float)y / rowHeight);
+        int i = (int)Math.floor(y / rowHeight);
 
         return getRowInsertionPoint(i, x);
     }
